requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
mcstatus==11.1.1
numpy==2.1.1
matplotlib==3.9.2
